move intersect to helpers
